% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4subpharma-package.R
\docType{package}
\name{r4subpharma-package}
\alias{r4subpharma}
\alias{r4subpharma-package}
\title{r4subpharma: 'pharmaverse' Adapters for R4SUB Submission Readiness Evidence}
\description{
Bridges the 'pharmaverse' clinical reporting stack and the R4SUB (Ready for Submission) ecosystem. Converts 'metacore' metadata objects and ADaM (Analysis Data Model) datasets - such as those built with 'admiral' - into standardized R4SUB evidence table rows via 'r4subcore', so that submission readiness can be scored with 'r4subscore' without changing an existing pharmaverse pipeline.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r4sub.github.io/r4subpharma/}
  \item \url{https://github.com/R4SUB/r4subpharma}
  \item Report bugs at \url{https://github.com/R4SUB/r4subpharma/issues}
}

}
\author{
\strong{Maintainer}: Pawan Rama Mali \email{prm@outlook.in} (\href{https://orcid.org/0000-0001-7864-5819}{ORCID}) [copyright holder]

}
\keyword{internal}
